requirements.txt
fastapi
sqlalchemy
pyjwt
bcrypt